move terminal to a panel
