+++ frontend/src/views/CodingView.vue
@@ -23,13 +23,7 @@
                   <web-debugger :key='debuggerPath' :file-path='debuggerPath'></web-debugger>
                 </el-tab-pane>
                 <el-tab-pane label="Terminal" name="second">
-                  <div>
-                    <el-tabs v-model="editableTabsValue" type="card" editable class="demo-tabs" @edit="handleTabsEdit" tab-position="top">
-                      <el-tab-pane v-for="item in editableTabs" :key="item.name" :label="item.title" :name="item.name">
-                        <component :is=item.content></component>
-                      </el-tab-pane>
-                    </el-tabs>
-                  </div>
+                  <terminal-panel></terminal-panel>
                 </el-tab-pane>
               </el-tabs>
             </div>
@@ -53,6 +47,7 @@ import {
 import FileSet from '../components/fileSet.vue'
 import webDebugger from '../components/webDebugger.vue'
 import webTerminal from '../components/webTerminal.vue'
+import TerminalPanel from '../components/TerminalPanel.vue'
 import { ElNotification } from 'element-plus'
 import type { TabsPaneContext } from 'element-plus'
 
@@ -149,52 +144,6 @@ const handleClick = (tab: TabsPaneContext, event: Event) => {
   console.log(tab, event)
 }
 
-// code for multi terminal
-let tabIndex = 2
-const editableTabsValue = ref('1')
-const editableTabs = ref([
-  {
-    title: 'Terminal 1',
-    name: '1',
-    // eslint-disable-next-line
-    content: webTerminal
-  },
-  {
-    title: 'Terminal 2',
-    name: '2',
-    // eslint-disable-next-line
-    content: webTerminal
-  }
-])
-
-const handleTabsEdit = (targetName: string, action: 'remove' | 'add') => {
-  if (action === 'add') {
-    const newTabName = `${++tabIndex}`
-    editableTabs.value.push({
-      title: 'Terminal ' + tabIndex,
-      name: newTabName,
-      content: webTerminal
-    })
-    editableTabsValue.value = newTabName
-  } else if (action === 'remove') {
-    const tabs = editableTabs.value
-    let activeName = editableTabsValue.value
-    if (activeName === targetName) {
-      tabs.forEach((tab, index) => {
-        if (tab.name === targetName) {
-          const nextTab = tabs[index + 1] || tabs[index - 1]
-          if (nextTab) {
-            activeName = nextTab.name
-          }
-        }
-      })
-    }
-
-    editableTabsValue.value = activeName
-    editableTabs.value = tabs.filter((tab) => tab.name !== targetName)
-  }
-}
-
 </script>
 
 <style scoped>
@@ -254,10 +203,5 @@ div {
 .btnArea .el-button {
   margin: 5px 2px;
 }
-.demo-tabs > .el-tabs__content {
-  padding: 32px;
-  color: #6b778c;
-  font-size: 32px;
-  font-weight: 600;
-}
+
 </style>
